requirements.txt
pandas
requests
SQLAlchemy
psycopg2-binary
streamlit
pyyaml
